eye_project-main 2/requirements_jetson.txt
# Smart Eye Diagnosis - Jetson (PyTorch)
Flask>=2.2,<4.0
Flask-Cors>=4.0,<6.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
numpy>=1.24,<3.0
opencv-python>=4.8,<5.0
Pillow>=10.0,<12.0
qrcode[pil]>=7.4,<9.0
reportlab>=4.0,<5.0
ultralytics>=8.0,<9.0
grad-cam>=1.5.0
# Use Jetson vendor wheel when possible
torch>=2.0
torchvision>=0.15
